refactor: model package에서 entity와 dto package 구분
