+++ src/main/java/lotto/domain/entity/Percent.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.entity;
 
 public class Percent {
