+++ src/main/java/lotto/domain/entity/Rankings.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.entity;
 
 import static lotto.configuration.IntegerConstants.COST_OF_LOTTO;
