+++ src/main/java/lotto/domain/entity/WinningResult.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.entity;
 
 import java.util.List;
