+++ src/test/java/lotto/domain/LottoTest.java
@@ -5,11 +5,8 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
-import lotto.util.IntegerListUtil;
-import lotto.util.StringUtil;
+import lotto.domain.entity.Lotto;
 import org.junit.jupiter.api.Test;
 
 public class LottoTest {
