+++ src/test/java/lotto/domain/ThousandUnitMoneyTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import lotto.domain.entity.ThousandUnitMoney;
 import org.junit.jupiter.api.Test;
 
 public class ThousandUnitMoneyTest {
